BJ-965: remove @RequestScope by reason of @Async request handling
refactor ThreeDsServerClient
