--- src/main/java/com/rbkmoney/threeds/server/storage/exception/RestClientThreeDsServerRequestException.java
@@ -1,10 +1,0 @@
-package com.rbkmoney.threeds.server.storage.exception;
-
-public class RestClientThreeDsServerRequestException extends RuntimeException {
-    public RestClientThreeDsServerRequestException() {
-    }
-
-    public RestClientThreeDsServerRequestException(String message) {
-        super(message);
-    }
-}
