--- src/main/java/com/rbkmoney/threeds/server/storage/exception/ThreeDsServerInternalException.java
@@ -1,10 +1,0 @@
-package com.rbkmoney.threeds.server.storage.exception;
-
-public class ThreeDsServerInternalException extends RuntimeException {
-    public ThreeDsServerInternalException() {
-    }
-
-    public ThreeDsServerInternalException(String message) {
-        super(message);
-    }
-}
