+++ src/test/java/com/rbkmoney/threeds/server/storage/config/EnableAutoConfigurationTestConfig.java
@@ -1,0 +1,21 @@
+package com.rbkmoney.threeds.server.storage.config;
+
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.flyway.FlywayAutoConfiguration;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.boot.autoconfigure.jdbc.JdbcTemplateAutoConfiguration;
+import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
+import org.springframework.boot.test.context.TestConfiguration;
+
+@TestConfiguration
+@EnableAutoConfiguration(
+        exclude = {
+                DataSourceAutoConfiguration.class,
+                JdbcTemplateAutoConfiguration.class,
+                HibernateJpaAutoConfiguration.class,
+                FlywayAutoConfiguration.class,
+        }
+)
+public class EnableAutoConfigurationTestConfig {
+
+}
