+++ src/test/java/com/rbkmoney/threeds/server/storage/handler/HandlerTest.java
@@ -2,34 +2,15 @@ package com.rbkmoney.threeds.server.storage.handler;
 
 import com.rbkmoney.damsel.three_ds_server_storage.*;
 import com.rbkmoney.geck.common.util.TypeUtil;
-import com.rbkmoney.threeds.server.storage.ThreeDsServerStorageApplication;
-import com.rbkmoney.threeds.server.storage.config.TestConfig;
+import com.rbkmoney.threeds.server.storage.config.AbstractConfigWithoutDao;
 import com.rbkmoney.threeds.server.storage.entity.CardRangeEntity;
 import com.rbkmoney.threeds.server.storage.entity.CardRangePk;
 import com.rbkmoney.threeds.server.storage.entity.ChallengeFlowTransactionInfoEntity;
 import com.rbkmoney.threeds.server.storage.entity.LastUpdatedEntity;
-import com.rbkmoney.threeds.server.storage.repository.CardRangeRepository;
-import com.rbkmoney.threeds.server.storage.repository.ChallengeFlowTransactionInfoRepository;
-import com.rbkmoney.threeds.server.storage.repository.LastUpdatedRepository;
-import com.rbkmoney.threeds.server.storage.repository.SerialNumRepository;
 import com.rbkmoney.woody.thrift.impl.http.THSpawnClientBuilder;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.flyway.FlywayAutoConfiguration;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.boot.autoconfigure.jdbc.JdbcTemplateAutoConfiguration;
-import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
-import org.springframework.boot.test.context.ConfigFileApplicationContextInitializer;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import java.net.URI;
 import java.time.LocalDateTime;
@@ -40,42 +21,13 @@ import java.util.UUID;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.when;
-import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest(
-        classes = {ThreeDsServerStorageApplication.class, TestConfig.class},
-        webEnvironment = RANDOM_PORT,
-        properties = {"spring.main.allow-bean-definition-overriding=true"})
-@EnableAutoConfiguration(
-        exclude = {
-                DataSourceAutoConfiguration.class,
-                JdbcTemplateAutoConfiguration.class,
-                HibernateJpaAutoConfiguration.class,
-                FlywayAutoConfiguration.class,
-        }
-)
-@TestPropertySource("classpath:application.yml")
-@ContextConfiguration(initializers = HandlerTest.Initializer.class)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
-public class HandlerTest {
+public class HandlerTest extends AbstractConfigWithoutDao {
 
-    @MockBean
-    private ChallengeFlowTransactionInfoRepository challengeFlowTransactionInfoRepository;
-
-    @MockBean
-    private CardRangeRepository cardRangeRepository;
-
-    @MockBean
-    private SerialNumRepository serialNumRepository;
-
-    @MockBean
-    private LastUpdatedRepository lastUpdatedRepository;
+    private static final int TIMEOUT = 555000;
 
     @Value("${local.server.port}")
     private int port;
-
-    private static final int TIMEOUT = 555000;
 
     private CardRangesStorageSrv.Iface cardRangesStorageClient;
     private ChallengeFlowTransactionInfoStorageSrv.Iface challengeFlowTransactionInfoStorageClient;
@@ -175,13 +127,5 @@ public class HandlerTest {
 
         ChallengeFlowTransactionInfo challengeFlowTransactionInfo = challengeFlowTransactionInfoStorageClient.getChallengeFlowTransactionInfo(transactionId);
         assertEquals(transactionId, challengeFlowTransactionInfo.getTransactionId());
-    }
-
-    public static class Initializer extends ConfigFileApplicationContextInitializer {
-
-        @Override
-        public void initialize(ConfigurableApplicationContext configurableApplicationContext) {
-            super.initialize(configurableApplicationContext);
-        }
     }
 }
