+++ src/test/java/com/rbkmoney/threeds/server/storage/repository/CardRangesRepositoryTest.java
@@ -1,5 +1,6 @@
 package com.rbkmoney.threeds.server.storage.repository;
 
+import com.rbkmoney.threeds.server.storage.config.AbstractDaoConfig;
 import com.rbkmoney.threeds.server.storage.entity.CardRangeEntity;
 import com.rbkmoney.threeds.server.storage.entity.CardRangePk;
 import org.junit.Before;
@@ -14,7 +15,7 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @RunWith(SpringRunner.class)
-public class CardRangesRepositoryTest extends PostgresRepositoryTest {
+public class CardRangesRepositoryTest extends AbstractDaoConfig {
 
     @Autowired
     private CardRangeRepository repository;
