+++ src/test/java/com/rbkmoney/threeds/server/storage/repository/ChallengeFlowTransactionInfoRepositoryTest.java
@@ -1,5 +1,6 @@
 package com.rbkmoney.threeds.server.storage.repository;
 
+import com.rbkmoney.threeds.server.storage.config.AbstractDaoConfig;
 import com.rbkmoney.threeds.server.storage.entity.ChallengeFlowTransactionInfoEntity;
 import org.junit.Before;
 import org.junit.Test;
@@ -15,7 +16,7 @@ import static org.junit.Assert.assertTrue;
 
 
 @RunWith(SpringRunner.class)
-public class ChallengeFlowTransactionInfoRepositoryTest extends PostgresRepositoryTest {
+public class ChallengeFlowTransactionInfoRepositoryTest extends AbstractDaoConfig {
 
     @Autowired
     private ChallengeFlowTransactionInfoRepository repository;
