+++ src/test/java/com/rbkmoney/threeds/server/storage/repository/LastUpdatedRepositoryTest.java
@@ -1,5 +1,6 @@
 package com.rbkmoney.threeds.server.storage.repository;
 
+import com.rbkmoney.threeds.server.storage.config.AbstractDaoConfig;
 import com.rbkmoney.threeds.server.storage.entity.LastUpdatedEntity;
 import com.rbkmoney.threeds.server.storage.mapper.LastUpdatedMapper;
 import org.junit.Before;
@@ -16,7 +17,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 @RunWith(SpringRunner.class)
-public class LastUpdatedRepositoryTest extends PostgresRepositoryTest {
+public class LastUpdatedRepositoryTest extends AbstractDaoConfig {
 
     @Autowired
     private LastUpdatedRepository repository;
